\name{fra}
\alias{fra}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
FRACTION for number
}
\description{
to turn numeric number into fraction form
}
\usage{
fra(x, j = 7)
}
\arguments{
  \item{x}{
a numeric number
}
  \item{j}{
Decimal digits default is 7
}
}
\value{
Return a charerter which shows the fraction equals x, x is a number
}
\references{
Friedrich Leisch, 2008 Creating R packages: A Tutorial
Zhang Jinlong, How to create R package under Windows
}
\author{
OuYang Ming
}
\examples{
x=1/3
fra(x)
}
\keyword{ caculate }
